backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9
google-auth==2.23.4
requests==2.31.0

